+++ src/v1/types/responses/service_info/builder.rs
@@ -5,10 +5,10 @@ use chrono::Utc;
 use url::Url;
 
 use crate::v1::types::responses::ServiceInfo;
-use crate::v1::types::responses::service::Artifact;
-use crate::v1::types::responses::service::Organization;
-use crate::v1::types::responses::service::ServiceType;
-use crate::v1::types::responses::service::TES_VERSION;
+use crate::v1::types::responses::service_info::Artifact;
+use crate::v1::types::responses::service_info::Organization;
+use crate::v1::types::responses::service_info::ServiceType;
+use crate::v1::types::responses::service_info::TES_VERSION;
 
 /// The default group to use for the service.
 pub const DEFAULT_GROUP: &str = "org.ga4gh";